+++ auto_route/lib/src/route/auto_route_config.dart
@@ -12,8 +12,7 @@ typedef AutoRoutePageBuilder = Widget Function(RouteData data);
 typedef TitleBuilder = String Function(BuildContext context, RouteData data);
 
 /// Signature for a function that builds a widget with [RouteData]
-typedef WidgetBuilderWithData = Widget Function(
-    BuildContext context, RouteData data);
+typedef WidgetBuilderWithData = Widget Function(BuildContext context, RouteData data);
 
 /// Signature for a function that builds the page [restorationId]
 /// Used in [AutoRoutePage]
@@ -116,9 +115,7 @@ class AutoRoute {
     List<AutoRoute>? children,
   })  : _path = path,
         _pageBuilder = page.builder,
-        _children = children != null && children.isNotEmpty
-            ? RouteCollection.fromList(children)
-            : null;
+        _children = children != null && children.isNotEmpty ? RouteCollection.fromList(children) : null;
 
   AutoRoute._change({
     required this.page,
@@ -268,9 +265,7 @@ class AutoRoute {
       meta: meta ?? this.meta,
       maintainState: maintainState ?? this.maintainState,
       fullscreenDialog: fullscreenDialog ?? this.fullscreenDialog,
-      children: children != null
-          ? (children.isEmpty ? null : RouteCollection.fromList(children))
-          : this.children,
+      children: children != null ? (children.isEmpty ? null : RouteCollection.fromList(children)) : this.children,
       //copy
       title: title ?? this.title,
       restorationId: restorationId ?? this.restorationId,
@@ -329,8 +324,7 @@ class MaterialRoute extends AutoRoute {
   }) : super._(
           type: RouteType.material(
             enablePredictiveBackGesture: enablePredictiveBackGesture,
-            predictiveBackPageTransitionsBuilder:
-                predictiveBackPageTransitionsBuilder,
+            predictiveBackPageTransitionsBuilder: predictiveBackPageTransitionsBuilder,
           ),
         );
 }
@@ -383,8 +377,7 @@ class AdaptiveRoute<R> extends AutoRoute {
           type: RouteType.adaptive(
             opaque: opaque,
             enablePredictiveBackGesture: enablePredictiveBackGesture,
-            predictiveBackPageTransitionsBuilder:
-                predictiveBackPageTransitionsBuilder,
+            predictiveBackPageTransitionsBuilder: predictiveBackPageTransitionsBuilder,
           ),
         );
 }
@@ -409,10 +402,8 @@ class CustomRoute<R> extends AutoRoute {
     super.allowSnapshotting = true,
     RouteTransitionsBuilder? transitionsBuilder,
     CustomRouteBuilder? customRouteBuilder,
-    @Deprecated('Use duration instead')
-    int? durationInMilliseconds,
-    @Deprecated('Use reverseDuration instead')
-    int? reverseDurationInMilliseconds,
+    @Deprecated('Use duration instead') int? durationInMilliseconds,
+    @Deprecated('Use reverseDuration instead') int? reverseDurationInMilliseconds,
     Duration? duration,
     Duration? reverseDuration,
     bool opaque = true,
@@ -422,25 +413,27 @@ class CustomRoute<R> extends AutoRoute {
     Color? barrierColor,
     bool enablePredictiveBackGesture = false,
     RouteTransitionsBuilder? predictiveBackPageTransitionsBuilder,
-  }) : super._(
+  })  : assert(durationInMilliseconds == null || duration == null, 'Use duration instead of durationInMilliseconds'),
+        assert(reverseDurationInMilliseconds == null || reverseDuration == null,
+            'Use reverseDuration instead of reverseDurationInMilliseconds'),
+        super._(
           type: RouteType.custom(
             transitionsBuilder: transitionsBuilder,
             customRouteBuilder: customRouteBuilder,
-            duration: duration ??
-                (durationInMilliseconds != null
-                    ? Duration(milliseconds: durationInMilliseconds)
-                    : null),
-            reverseDuration: reverseDuration ??
-                (reverseDurationInMilliseconds != null
-                    ? Duration(milliseconds: reverseDurationInMilliseconds)
-                    : null),
+            duration: duration,
+            reverseDuration: reverseDuration,
+            // todo: remove deprecated members
+            // ignore: deprecated_member_use_from_same_package
+            durationInMilliseconds: durationInMilliseconds,
+            // todo: remove deprecated members
+            // ignore: deprecated_member_use_from_same_package
+            reverseDurationInMilliseconds: reverseDurationInMilliseconds,
             opaque: opaque,
             barrierDismissible: barrierDismissible,
             barrierLabel: barrierLabel,
             barrierColor: barrierColor,
             enablePredictiveBackGesture: enablePredictiveBackGesture,
-            predictiveBackPageTransitionsBuilder:
-                predictiveBackPageTransitionsBuilder,
+            predictiveBackPageTransitionsBuilder: predictiveBackPageTransitionsBuilder,
           ),
         );
 }
@@ -562,8 +555,7 @@ class RouteCollection {
   ///
   /// if this [RouteCollection] is created by the router [root] will be true
   /// else if it's created by a parent route-entry it will be false
-  factory RouteCollection.fromList(List<AutoRoute> routes,
-      {bool root = false}) {
+  factory RouteCollection.fromList(List<AutoRoute> routes, {bool root = false}) {
     final routesMarkedInitial = routes.where((e) => e.initial);
     throwIf(routesMarkedInitial.length > 1,
         'Invalid data\nThere are more than one initial route in this collection\n${routesMarkedInitial.map((e) => e.name)}');